+++ packages/framework/src/Form/LocalizedFullWidthType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shopsys\FrameworkBundle\Form;
+
+use Shopsys\FrameworkBundle\Form\Locale\LocalizedType;
+
+class LocalizedFullWidthType extends LocalizedType
+{
+}